stop wiping tags whenever you manually patch an api endpoint
man that was dumb. why was i doing that?

+++ src/challenge/serializers.py
@@ -136,10 +136,11 @@ class CreateChallengeSerializer(serializers.ModelSerializer):
         return challenge
 
     def update(self, instance, validated_data):
-        tags = validated_data.pop('tags', [])
-        Tag.objects.filter(challenge=instance).delete()
-        for tag_data in tags:
-            Tag.objects.create(challenge=instance, **tag_data)
+        tags = validated_data.pop('tags', None)
+        if tags:
+            Tag.objects.filter(challenge=instance).delete()
+            for tag_data in tags:
+                Tag.objects.create(challenge=instance, **tag_data)
         return super().update(instance, validated_data)
